# Minor fingerprint improvements