--- a/src/lsm/bloom_filter.zig
+++ b/src/lsm/bloom_filter.zig
@@ -11,7 +11,7 @@ pub const Fingerprint = struct {
     /// Hash value used to map key to block.
     hash: u32,
     /// Mask of bits set in the block for the key.
-    mask: meta.Vector(8, u32),
+    mask: @Vector(8, u32),
 
     pub fn create(hash: u64) Fingerprint {
         const hash_lower = @truncate(u32, hash);
@@ -20,7 +20,7 @@ pub const Fingerprint = struct {
         // TODO These constants are from the paper and we understand them to be arbitrary odd
         // integers. Experimentally compare the performance of these with other randomly chosen
         // odd integers to verify/improve our understanding.
-        const odd_integers: meta.Vector(8, u32) = [8]u32{
+        const odd_integers: @Vector(8, u32) = [8]u32{
             0x47b6137b,
             0x44974d91,
             0x8824ad5b,
@@ -36,7 +36,7 @@ pub const Fingerprint = struct {
 
         return .{
             .hash = hash_upper,
-            .mask = @splat(8, @as(u32, 1)) << @intCast(meta.Vector(8, u5), bit_indexes),
+            .mask = @splat(8, @as(u32, 1)) << @intCast(@Vector(8, u5), bit_indexes),
         };
     }
 };
@@ -44,37 +44,37 @@ pub const Fingerprint = struct {
 /// Add the key with the given fingerprint to the filter.
 /// filter.len must be a multiple of 32.
 pub fn add(fingerprint: Fingerprint, filter: []u8) void {
-    comptime assert(@sizeOf(meta.Vector(8, u32)) == 32);
+    comptime assert(@sizeOf(@Vector(8, u32)) == 32);
 
     assert(filter.len > 0);
-    assert(filter.len % @sizeOf(meta.Vector(8, u32)) == 0);
+    assert(filter.len % @sizeOf(@Vector(8, u32)) == 0);
 
     const blocks = mem.bytesAsSlice([8]u32, filter);
     const index = block_index(fingerprint.hash, filter.len);
 
-    const current: meta.Vector(8, u32) = blocks[index];
+    const current: @Vector(8, u32) = blocks[index];
     blocks[index] = current | fingerprint.mask;
 }
 
 /// Check if the key with the given fingerprint may have been added to the filter.
 /// filter.len must be a multiple of 32.
 pub fn may_contain(fingerprint: Fingerprint, filter: []const u8) bool {
-    comptime assert(@sizeOf(meta.Vector(8, u32)) == 32);
+    comptime assert(@sizeOf(@Vector(8, u32)) == 32);
 
     assert(filter.len > 0);
-    assert(filter.len % @sizeOf(meta.Vector(8, u32)) == 0);
+    assert(filter.len % @sizeOf(@Vector(8, u32)) == 0);
 
     const blocks = mem.bytesAsSlice([8]u32, filter);
     const index = block_index(fingerprint.hash, filter.len);
 
-    const current: meta.Vector(8, u32) = blocks[index];
+    const current: @Vector(8, u32) = blocks[index];
     return @reduce(.Or, ~current & fingerprint.mask) == 0;
 }
 
 inline fn block_index(hash: u32, size: usize) u32 {
     assert(size > 0);
 
-    const block_count = @divExact(size, @sizeOf(meta.Vector(8, u32)));
+    const block_count = @divExact(size, @sizeOf(@Vector(8, u32)));
     return @intCast(u32, (@as(u64, hash) * block_count) >> 32);
 }