--- a/src/lsm/groove.zig
+++ b/src/lsm/groove.zig
@@ -12,9 +12,11 @@ const TreeType = @import("tree.zig").TreeType;
 const GridType = @import("grid.zig").GridType;
 const CompositeKey = @import("composite_key.zig").CompositeKey;
 const NodePool = @import("node_pool.zig").NodePool(constants.lsm_manifest_node_size, 16);
+const Fingerprint = @import("bloom_filter.zig").Fingerprint;
 
 const snapshot_latest = @import("tree.zig").snapshot_latest;
 const compaction_snapshot_for_op = @import("tree.zig").compaction_snapshot_for_op;
+const key_fingerprint = @import("tree.zig").key_fingerprint;
 
 fn ObjectTreeHelpers(comptime Object: type) type {
     assert(@hasField(Object, "timestamp"));
@@ -387,12 +389,16 @@ pub fn GrooveType(
 
         const primary_field = if (has_id) "id" else "timestamp";
         const PrimaryKey = @TypeOf(@field(@as(Object, undefined), primary_field));
-        const PrefetchKey = struct { key: union(enum) {
-            id: PrimaryKey,
-            timestamp: u64,
-        }, level: u8 };
+        const PrefetchKey = struct {
+            key: union(enum) {
+                id: PrimaryKey,
+                timestamp: u64,
+            },
+            fingerprint: Fingerprint,
+            level: u8,
+        };
 
-        const PrefetchKeys = std.AutoHashMapUnmanaged(PrefetchKey, void);
+        const PrefetchKeys = std.AutoArrayHashMapUnmanaged(PrefetchKey, void);
 
         const PrefetchObjectsContext = struct {
             pub inline fn hash(_: PrefetchObjectsContext, object: Object) u64 {
@@ -582,7 +588,10 @@ pub fn GrooveType(
         /// This must be called by the state machine for every key to be prefetched.
         /// We tolerate duplicate IDs enqueued by the state machine.
         /// For example, if all unique operations require the same two dependencies.
-        pub fn prefetch_enqueue(groove: *Groove, key: PrimaryKey) void {
+        pub fn prefetch_enqueue(
+            groove: *Groove,
+            key: PrimaryKey,
+        ) void {
             if (has_id) {
                 if (!groove.ids.key_range_contains(groove.prefetch_snapshot.?, key)) return;
                 groove.prefetch_from_memory_by_id(key);
@@ -597,7 +606,8 @@ pub fn GrooveType(
         /// TODO: We may have to remove this function once Fed's prefetching changes are merged,
         /// since those changes remove lookup_from_memory.
         fn prefetch_from_memory_by_id(groove: *Groove, id: PrimaryKey) void {
-            switch (groove.ids.lookup_from_memory(groove.prefetch_snapshot.?, id)) {
+            const fingerprint = key_fingerprint(id);
+            switch (groove.ids.lookup_from_memory(groove.prefetch_snapshot.?, id, fingerprint)) {
                 .negative => {},
                 .positive => |id_tree_value| {
                     if (IdTreeValue.tombstone(id_tree_value)) return;
@@ -607,6 +617,7 @@ pub fn GrooveType(
                     groove.prefetch_keys.putAssumeCapacity(.{
                         .key = .{ .id = id },
                         .level = level,
+                        .fingerprint = fingerprint,
                     }, {});
                 },
             }
@@ -617,7 +628,12 @@ pub fn GrooveType(
         /// TODO: We may have to remove this function once Fed's prefetching changes are merged,
         /// since those changes remove lookup_from_memory.
         fn prefetch_from_memory_by_timestamp(groove: *Groove, timestamp: u64) void {
-            switch (groove.objects.lookup_from_memory(groove.prefetch_snapshot.?, timestamp)) {
+            const fingerprint = key_fingerprint(timestamp);
+            switch (groove.objects.lookup_from_memory(
+                groove.prefetch_snapshot.?,
+                timestamp,
+                fingerprint,
+            )) {
                 .negative => {},
                 .positive => |object| {
                     assert(!ObjectTreeHelpers(Object).tombstone(object));
@@ -626,6 +642,7 @@ pub fn GrooveType(
                 .possible => |level| {
                     groove.prefetch_keys.putAssumeCapacity(.{
                         .key = .{ .timestamp = timestamp },
+                        .fingerprint = fingerprint,
                         .level = level,
                     }, {});
                 },
@@ -643,7 +660,7 @@ pub fn GrooveType(
                 .groove = groove,
                 .callback = callback,
                 .snapshot = groove.prefetch_snapshot.?,
-                .key_iterator = groove.prefetch_keys.keyIterator(),
+                .key_iterator = groove.prefetch_keys.iterator(),
             };
             groove.prefetch_snapshot = null;
             context.start_workers();
@@ -654,8 +671,7 @@ pub fn GrooveType(
             callback: *const fn (*PrefetchContext) void,
             snapshot: u64,
 
-            key_iterator: PrefetchKeys.KeyIterator,
-
+            key_iterator: PrefetchKeys.Iterator,
             /// The goal is to fully utilize the disk I/O to ensure the prefetch completes as
             /// quickly as possible, so we run multiple lookups in parallel based on the max
             /// I/O depth of the Grid.
@@ -731,9 +747,12 @@ pub fn GrooveType(
             lookup: LookupContext = undefined,
 
             fn lookup_start_next(worker: *PrefetchWorker) void {
-                const prefetch_key = worker.context.key_iterator.next() orelse {
-                    worker.context.worker_finished();
-                    return;
+                const prefetch_key: *const PrefetchKey = blk: {
+                    const entry = worker.context.key_iterator.next() orelse {
+                        worker.context.worker_finished();
+                        return;
+                    };
+                    break :blk entry.key_ptr;
                 };
 
                 // prefetch_enqueue() ensures that the tree's cache is checked before queueing the
@@ -747,6 +766,7 @@ pub fn GrooveType(
                                 .context = worker.lookup.get(.id),
                                 .snapshot = worker.context.snapshot,
                                 .key = id,
+                                .fingerprint = prefetch_key.fingerprint,
                                 .level_min = prefetch_key.level,
                             });
                         } else unreachable;
@@ -757,6 +777,7 @@ pub fn GrooveType(
                             .context = worker.lookup.get(.object),
                             .snapshot = worker.context.snapshot,
                             .key = timestamp,
+                            .fingerprint = prefetch_key.fingerprint,
                             .level_min = prefetch_key.level,
                         });
                     },
@@ -781,9 +802,11 @@ pub fn GrooveType(
             }
 
             fn lookup_by_timestamp(worker: *PrefetchWorker, timestamp: u64) void {
+                const fingerprint = key_fingerprint(timestamp);
                 switch (worker.context.groove.objects.lookup_from_memory(
                     worker.context.snapshot,
                     timestamp,
+                    fingerprint,
                 )) {
                     .negative => {
                         lookup_object_callback(worker.lookup.get(.object), null);
@@ -797,6 +820,7 @@ pub fn GrooveType(
                             .context = worker.lookup.get(.object),
                             .snapshot = worker.context.snapshot,
                             .key = timestamp,
+                            .fingerprint = fingerprint,
                             .level_min = level_min,
                         });
                     },