--- a/src/lsm/tree_fuzz.zig
+++ b/src/lsm/tree_fuzz.zig
@@ -22,6 +22,7 @@ const allocate_block = @import("grid.zig").allocate_block;
 const NodePool = @import("node_pool.zig").NodePool(constants.lsm_manifest_node_size, 16);
 const TableUsage = @import("table.zig").TableUsage;
 const TableType = @import("table.zig").TableType;
+const key_fingerprint = @import("tree.zig").key_fingerprint;
 
 const Grid = GridType(Storage);
 const SuperBlock = vsr.SuperBlockType(Storage);
@@ -280,9 +281,10 @@ fn EnvironmentType(comptime table_usage: TableUsage) type {
 
         pub fn get(env: *Environment, key: Key) ?*const Key.Value {
             env.change_state(.fuzzing, .tree_lookup);
-
             env.lookup_value = null;
-            switch (env.tree.lookup_from_memory(snapshot_latest, key)) {
+
+            const fingerprint = key_fingerprint(key);
+            switch (env.tree.lookup_from_memory(snapshot_latest, key, fingerprint)) {
                 .negative => {
                     get_callback(&env.lookup_context, null);
                 },
@@ -295,6 +297,7 @@ fn EnvironmentType(comptime table_usage: TableUsage) type {
                         .context = &env.lookup_context,
                         .snapshot = env.tree.lookup_snapshot_max.?,
                         .key = key,
+                        .fingerprint = fingerprint,
                         .level_min = level_min,
                     });
                 },