- Change the fingerprint for composite keys, taking only the prefix, not the entire key.
Rationale: We don't use composite keys for lookups. However, the bloom filter can be useful for scans with equality conditions by prefix.

- Propagate and store the key's fingerprint during the prefetch path.
Rationale: Hashing the key only once.

- Change PrefetchKeys' implementation from HashTable to ArrayList.
Rationale: It's used only for iterating so that we can avoid the costs of the hashtable iterator.

A small but consistent performance improvement can be noticed:

Before:
```
1225 batches in 80.72 s
load offered = 1000000 tx/s
load accepted = 123877 tx/s
batch latency p00 = 0 ms
batch latency p10 = 17 ms
batch latency p20 = 20 ms
batch latency p30 = 25 ms
batch latency p40 = 26 ms
batch latency p50 = 27 ms
batch latency p60 = 28 ms
batch latency p70 = 29 ms
batch latency p80 = 31 ms
batch latency p90 = 32 ms
batch latency p100 = 4076 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3787 ms
transfer latency p20 = 7818 ms
transfer latency p30 = 14363 ms
transfer latency p40 = 21452 ms
transfer latency p50 = 29801 ms
transfer latency p60 = 37824 ms
transfer latency p70 = 46693 ms
transfer latency p80 = 54293 ms
transfer latency p90 = 63082 ms
transfer latency p100 = 70743 ms
```

After:
```
1225 batches in 77.64 s
load offered = 1000000 tx/s
load accepted = 128797 tx/s
batch latency p00 = 0 ms
batch latency p10 = 17 ms
batch latency p20 = 18 ms
batch latency p30 = 20 ms
batch latency p40 = 26 ms
batch latency p50 = 26 ms
batch latency p60 = 28 ms
batch latency p70 = 29 ms
batch latency p80 = 30 ms
batch latency p90 = 31 ms
batch latency p100 = 3708 ms
transfer latency p00 = 0 ms
transfer latency p10 = 3700 ms
transfer latency p20 = 7259 ms
transfer latency p30 = 13829 ms
transfer latency p40 = 20573 ms
transfer latency p50 = 28181 ms
transfer latency p60 = 36190 ms
transfer latency p70 = 44167 ms
transfer latency p80 = 51612 ms
transfer latency p90 = 60137 ms
transfer latency p100 = 67658 ms
```